chore: Move LocationToMapBinder to map , Relocate LocationAutoMover to map
