+++ app/src/main/java/com/example/capstone_map/map/LocationAutoMover.java
@@ -1,11 +1,11 @@
-package com.example.capstone_map.navigation;
+package com.example.capstone_map.map;
 
 import android.content.Context;
 import android.location.Location;
 import android.util.Log;
 
-import com.example.capstone_map.navigation.LocationTracker;
-import com.example.capstone_map.navigation.LocationUpdateCallback;
+import com.example.capstone_map.location.LocationTracker;
+import com.example.capstone_map.location.LocationUpdateCallback;
 import com.skt.Tmap.TMapView;
 
 public class LocationAutoMover implements LocationUpdateCallback {
